+++ packages/polaris-viz-core/src/index.ts
@@ -113,6 +113,7 @@ export {
   usePolarisVizContext,
   useYScale,
   useUniqueId,
+  useAriaLabel,
 } from './hooks';
 export {
   Bar,