+++ packages/polaris-viz/src/components/HorizontalBarChart/tests/HorizontalBarChart.test.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import {mount} from '@shopify/react-testing';
+import {DataSeries} from '@shopify/polaris-viz-core';
 
 import {
   HorizontalBarChart,
@@ -13,7 +14,15 @@ jest.mock('@shopify/polaris-viz-core/src/utilities', () => ({
 }));
 
 const mockProps: HorizontalBarChartProps = {
-  data: [{name: 'Test', data: [{value: 10, key: 'data'}]}],
+  data: [
+    {
+      name: 'Test',
+      data: [
+        {value: 10, key: 'Label 01'},
+        {value: 12, key: 'Label 02'},
+      ],
+    },
+  ],
   xAxisOptions: {
     labelFormatter: (value) => `${value}`,
     hide: false,
@@ -35,5 +44,19 @@ describe('<HorizontalBarChart />', () => {
     const barChart = mount(<HorizontalBarChart {...mockProps} />);
 
     expect(barChart).toContainReactComponent(Chart);
+  });
+
+  it('correctly rerenders if new data object is passed in as prop', () => {
+    const alteredData = [
+      {
+        name: 'Test',
+        data: [{value: 12, key: 'Label 02'}],
+      },
+    ] as DataSeries[];
+
+    const chart = mount(<HorizontalBarChart {...mockProps} />);
+    chart.setProps({...mockProps, data: alteredData});
+
+    expect(chart).toContainReactComponent(Chart);
   });
 });